[refactor] api response 위치 이동 #258
