+++ src/main/java/classfit/example/classfit/attendance/controller/AttendanceExportController.java
@@ -4,7 +4,7 @@ import classfit.example.classfit.attendance.controller.docs.AttendanceExportCont
 import classfit.example.classfit.attendance.dto.response.StudentAttendanceResponse;
 import classfit.example.classfit.attendance.service.AttendanceExportService;
 import classfit.example.classfit.auth.annotation.AuthMember;
-import classfit.example.classfit.common.CustomApiResponse;
+import classfit.example.classfit.common.response.CustomApiResponse;
 import classfit.example.classfit.member.domain.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.GetMapping;
