+++ src/main/java/classfit/example/classfit/auth/security/filter/CustomLoginFilter.java
@@ -3,7 +3,7 @@ package classfit.example.classfit.auth.security.filter;
 import classfit.example.classfit.auth.dto.request.UserRequest;
 import classfit.example.classfit.auth.security.custom.CustomAuthenticationToken;
 import classfit.example.classfit.auth.security.jwt.JWTUtil;
-import classfit.example.classfit.common.CustomApiResponse;
+import classfit.example.classfit.common.response.CustomApiResponse;
 import classfit.example.classfit.common.exception.ClassfitAuthException;
 import classfit.example.classfit.common.exception.ClassfitException;
 import classfit.example.classfit.common.util.CookieUtil;
