+++ src/main/java/classfit/example/classfit/calendarCategory/controller/docs/CalendarCategoryControllerDocs.java
@@ -6,7 +6,7 @@ import classfit.example.classfit.calendarCategory.dto.request.CalendarCategoryUp
 import classfit.example.classfit.calendarCategory.dto.response.CalendarCategoryCreateResponse;
 import classfit.example.classfit.calendarCategory.dto.response.CalendarCategoryListResponse;
 import classfit.example.classfit.calendarCategory.dto.response.CalendarCategoryResponse;
-import classfit.example.classfit.common.CustomApiResponse;
+import classfit.example.classfit.common.response.CustomApiResponse;
 import classfit.example.classfit.member.domain.Member;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
