+++ src/main/java/classfit/example/classfit/category/controller/docs/ClassInfoControllerDocs.java
@@ -2,7 +2,7 @@ package classfit.example.classfit.category.controller.docs;
 
 import classfit.example.classfit.auth.annotation.AuthMember;
 import classfit.example.classfit.category.dto.response.ClassInfoResponse;
-import classfit.example.classfit.common.CustomApiResponse;
+import classfit.example.classfit.common.response.CustomApiResponse;
 import classfit.example.classfit.member.domain.Member;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
