+++ src/main/java/classfit/example/classfit/category/controller/docs/SubClassControllerDocs.java
@@ -3,7 +3,7 @@ package classfit.example.classfit.category.controller.docs;
 import classfit.example.classfit.auth.annotation.AuthMember;
 import classfit.example.classfit.category.dto.request.SubClassRequest;
 import classfit.example.classfit.category.dto.response.SubClassResponse;
-import classfit.example.classfit.common.CustomApiResponse;
+import classfit.example.classfit.common.response.CustomApiResponse;
 import classfit.example.classfit.member.domain.Member;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
