+++ src/main/java/classfit/example/classfit/common/exception/ClassfitExceptionControllerAdivce.java
@@ -1,6 +1,6 @@
 package classfit.example.classfit.common.exception;
 
-import classfit.example.classfit.common.CustomApiResponse;
+import classfit.example.classfit.common.response.CustomApiResponse;
 import com.fasterxml.jackson.databind.exc.MismatchedInputException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
