+++ src/main/java/classfit/example/classfit/member/controller/MemberController.java
@@ -1,7 +1,7 @@
 package classfit.example.classfit.member.controller;
 
 import classfit.example.classfit.auth.annotation.AuthMember;
-import classfit.example.classfit.common.CustomApiResponse;
+import classfit.example.classfit.common.response.CustomApiResponse;
 import classfit.example.classfit.member.controller.docs.MemberControllerDocs;
 import classfit.example.classfit.member.domain.Member;
 import classfit.example.classfit.member.dto.request.MemberPasswordRequest;
