+++ src/main/java/classfit/example/classfit/scoreReport/controller/ScoreReportController.java
@@ -1,7 +1,7 @@
 package classfit.example.classfit.scoreReport.controller;
 
 import classfit.example.classfit.auth.annotation.AuthMember;
-import classfit.example.classfit.common.CustomApiResponse;
+import classfit.example.classfit.common.response.CustomApiResponse;
 import classfit.example.classfit.member.domain.Member;
 import classfit.example.classfit.scoreReport.controller.docs.ScoreReportControllerDocs;
 import classfit.example.classfit.scoreReport.dto.process.ReportExam;
