+++ src/main/java/classfit/example/classfit/student/controller/StudentController.java
@@ -1,7 +1,7 @@
 package classfit.example.classfit.student.controller;
 
 import classfit.example.classfit.auth.annotation.AuthMember;
-import classfit.example.classfit.common.CustomApiResponse;
+import classfit.example.classfit.common.response.CustomApiResponse;
 import classfit.example.classfit.member.domain.Member;
 import classfit.example.classfit.student.controller.docs.StudentControllerDocs;
 import classfit.example.classfit.student.dto.request.StudentRequest;
